requirements.txt
pymodbus
pyserial
pandas
matplotlib
jupyter
